Remove the .Normalize() to fix PointAtInternal
The Normalize method call when calculating the current vector is causing errors.  Line 166 does not want a normalized vector, it wants to compute how much of the full vector is needed to reach the desired paramter.  
Suggest adding a test in Polyline tests that looks like
``` C#
            var r = Polygon.Rectangle(2,2);
            Assert.Equal(new Vector3(1,1,0), r.PointAt(0.5));
```
This test, which uses a rectangle not of unit length will reveal the problem.