AddVerticesWithCustomExtension always insert vertices to the grid.
Before, segments where just inserted only if it was "inside" the grid - extended segment must have hit at lest one edge on both sides.
This is inconsistent with other cases of AddVertices and too confusing.
Now points are always added to the grid and connected.
The responsibilities for polyline correctness is shifted to the end user.